util: Local.Context better representation
Problem:

The Local.Context is using a mutable Array to hold all locals, and it does two
deep copies every time we invoke the let method.

Solution:

Local.Context becomes a Key-Value structured type, it stores key as Local.Key
and value as Option[Any]. In this case, we can avoid wrapping the value in an
Option as Map does, and it avoid copying the entire array everytime as before.
This type has the better run-time performance for our use cases.

JIRA Issues: CSL-5667

Differential Revision: https://phabricator.twitter.biz/D182478