util-jvm: make catch exception handling less restrictive
Problem

When using the twitter util-jvm with a non-Hotpost JVM, the
library tries to create an interface that applies reflection
on Hotspot specific classes. This causes a Finagle application
to crash with the stacktrace outlined in #213.

Solution

The previous code catches NonFatal exceptions and returns a NilJvm
in that scenario. This PR changes this behavior so that a NilJvm
is returned on NonFatal and UnsatisfiedLinkedError exceptions when
a new Hotspot instance creation is attempted.

Signed-off-by: Bryce Anderson <banderson@twitter.com>

JIRA Issues: CSL-6666

Differential Revision: https://phabricator.twitter.biz/D190654