util: make joinWith short circuit on first failure
Problem

The promise returned by joinWith should reflect the failure of either future
immediately. We rely on this to express computations where we want to wait on
two futures, unless one of them fails, then we satisfy the promise with the
first failure. joinWith is supposed to do this, but it actually sequences the
futures, no different from a flatMap.

Solution

Change joinWith so that it waits for two futures and applies a function to
the two results. If either future fails, the returned promise must
reflect the failure immediately.

JIRA Issues: CSL-6632

Differential Revision: https://phabricator.twitter.biz/D191342