util: add capture point for Future.joinWith
Problem / Solution

Capture points refer to closures and help the debugger to stitch
together an asynchronous stack trace. Because of a recent change to
Future.joinWith, the closure referred to by the former capture point was
replaced with different closure. This change reinstates the capture
point, updated to correctly address the new closure.

JIRA Issues: CSL-6693

Differential Revision: https://phabricator.twitter.biz/D193156