util: Introduce c.t.io.Pipe, a primitive that syncs reader with writer
# Problem

Despite serving a cornerstone purpose, `Reader.writable` isn't really structured as its own concept (abstraction),
which yields an unnecessary confusion: is that //mostly// a reader or a writer? is there a `Writer.readable`?

# Solution

Promote `Reader.writable` to its own abstraction, `Pipe[A]` - an in-memory primitive that synchronizes a reader
with a writer. Pipes allow connecting a producer with a consumer, a code that requires `Writer` with a code that
operates on `Reader`s.

JIRA Issues: CSL-6752

Differential Revision: https://phabricator.twitter.biz/D199536