util-slf4j-api: Use marker information when determining if log level is enabled
Problem

Markers are not currently used in determining if a log level is enabled on
API methods which accept markers.

Solution

Ensure we pass the marker to the "isLevelEnabled" functions.

Result

More correct determination of if a log level is enabled for a method which
accepts marker data.

JIRA Issues: CSL-6837

Differential Revision: https://phabricator.twitter.biz/D202387