util-logging: New way to construct ScribeHandler for java interoperability
Problem

The ScribeHandler class currently does not have good interoperability with Java
while constructing the ScribeHandler object because of Scala's default
parameters to constructors/companion object apply method. Java when calling the
Scala's method needs to provide values to all those default arguments which is
not ideal especially if those default arguments are not surfaced to the client
building the ScribeHandler.

Solution

The solution is to have an easy way to create the ScribeHandler instance when it
only needs the category and formatter as the parameters. This is not meant to
have all the possible ways of creating this object as a solution (for example,
by using a Builder Pattern in the ScribeHandler class)

Result

With this change there will be an additional way to construct a ScribeHandler
instance - especially for use in Java.

JIRA Issues: CSL-6923

Differential Revision: https://phabricator.twitter.biz/D208928