util-core: Introduce Reader.framed
Problem

Network protocols often frame data based on a size field preceding the message
so that a reader knows how much data to read before decoding the message. For
instance, this is the case with the mux protocol.

Solution

Add a Reader[Buf] implementation that will frame Bufs based on a given
pattern.

JIRA Issues: CSL-6778

Differential Revision: https://phabricator.twitter.biz/D212396