util-tunable: Make tunables observable via Vars
Problem

Tunables are currently "pull" only, to get the value you need to read it from
the tunable.  There are cases however where it would be useful to be notified
when a tunable changes (eg, to update a configuration for APIs that don't
support tunables).

Solution

Add an `asVar` API to Tunable that returns a `Var` representing the tunable's
(current and future) value.

Differential Revision: https://phabricator.twitter.biz/D211622