twitter-oss: Update TravisCI to use sudo = required
Problem

It has been recommended by the TravisCI support that we switch our
builds to use `sudo: required` in order to get a fully virtualized
environment for running our builds.

See:
https://docs.travis-ci.com/user/reference/trusty/#fully-virtualized-via-sudo-required

We hadn't switched as we were beholden to needing a Java install which
wasn't readily available previously but is available in the TravisCI "trusty"
image.

Solution

Switch the TravisCI configuration to use `sudo: required` and `dist: trusty`.

Differential Revision: https://phabricator.twitter.biz/D218377