TravisCI: Remove Build Stages configuration
Problem

Time is a flat circle. Everything we've ever done or
will do, we're going to do over and over again.

Build stages rely on the caching of artifacts to S3
in order fro jobs to share state. While this seems
like a great idea in theory. The practice has proven
to be somewhat untenable.

Solution

Walk away from build stages. This really only affects
the Finagle build as it is the only build which remains
a matrix build and thus the only build which potentially
runs the dodo step multiple times. Given the complexity
around trying to get build stages to work and the
cache to be properly created and shared for different
scala versions, this is an acceptable amount of overhead.

JIRA Issues: CSL-7054

TBR=true

Differential Revision: https://phabricator.twitter.biz/D222286