util-core: Add contramap for Writer
Problem:
We want to make Writer API polymorphic regarding the type it operates on.

Solution:
Writer becomes a contravariant functor by adding `contramap`: if a function `f`
which transforms `B` into `A` is given, and the `Writer[A]` is implemented, we get the
`Writer[B]` by calling `contramap`.

JIRA Issues: CSL-6686

Differential Revision: https://phabricator.twitter.biz/D225686