util-core: test all AsyncStream ctors and fix tail for Embed
Problem

AsyncStream has several implementations, but not all of them are covered by
unit tests. The new tests revealed a small bug in the Embed case for tail.

The future returned by tail should always resolve to Future.None if the
AsyncStream is empty. This property aids clarity: None means there is nothing
left to process, and Some(x) means that there is. Specifically, in the Some(x)
case, x must not be empty. The bug in the Embed case is that it's possible to get
Some(x), where x is empty.

Solution

In the Embed case, we should wait for the embedded (within the future)
AsyncStream to materialize, then check if it's empty. And if the embedded
AsyncStream resolves to another embedded AsyncStream, we should recurse until
we can get our hands on a non-Embed AsyncStream.

Differential Revision: https://phabricator.twitter.biz/D241513