util-core: Make WaitQueue and derived types abstract classes
Problem

We use `trait`s as the foundations for data strutures used in our
`Promise` implementation and quite a few megamorphic methods. This
can require the JVM to perform invokeinterface calls, which in
contrast to invokevirtual calls, required searching through the list
of implemented interfaces to find the interface of interest. This
can result in meaningful overhead for very high throughput code
dealing in classes that implement many interfaces.

Solution

Restruture the `WaitQueue` class and its subtypes as abstract
classes.

Differential Revision: https://phabricator.twitter.biz/D242570