util-core: Pipe honor deadline on close
Problem

When closing a Pipe, the deadline is not honored. This is not strictly incorrect
but it is possible to try to allow reads to occur before the deadline.

Solution

Introduce an asynchronous close.

Result

Reads will succeed up until the close deadline. The default behavior will not be
affected, the close will only be async if the Pipe is instantiated with a
Timer.

JIRA Issues: CSL-3843

Differential Revision: https://phabricator.twitter.biz/D229728