util-app: Remove warning when registering App
Problem

In many cases, there is a single `c.t.app.App` instance is
registered in the singleton `c.t.app.App` object, however in
testing there may be cases where multiple instances are
started and thus registered. We log a warning but don't do
anything specifically to prevent multiple registrations as
the expectation is that the "last one wins". Thus the utility
of the warning is suspect.

Solution

Remove the warning log statement.

JIRA Issues: CSL-6034

Differential Revision: https://phabricator.twitter.biz/D273094