util-core: Don't allow AsyncSemaphore Permits to be released several times
Problem

It is possible to call release on the same permit twice, thus releasing the
permit and a waiter.

Solution

AsyncSemaphore permits should be idempotent so that if we call release twice on
a permit, it will only release the permit that is being released.

Result

With this change, you can call release on the same permit twice but it will only
release that one permit.

JIRA Issues: CSL-6358

Differential Revision: https://phabricator.twitter.biz/D274017