util-jvm: Make Hotspot.foreachGc more robust
Problem

The implemenation of `Hotspot.foreachGc` relies on the default
implementation from `Jvm`. This in turn relies on a mechanism that can
(mostly silently) miss garbage collection events.

Solution

Have a Hotspot-specific implementation that relies on
`GarbageCollectionNotificationInfo`, in a similar fashion to what is
done for tracking allocations.

Result

`Hotspot.foreachGc` is more robust in that it should no longer miss
garbage collection events.

JIRA Issues: CSL-7663

Differential Revision: https://phabricator.twitter.biz/D281962