util-logging: Add histogram stats of batch size to scribe handler
Problem

The `c.t.logging.ScribeHandler` does not provide stats on the size of
the batches sent to Scribe despite this being critical information for debugging
issues in scribe servers.

Solution

Add histogram stats for the size of batches sent to Scribe servers
(both number of events and total byte count).

Result

The size of batches sent to Scribe is visible.

JIRA Issues: PUBSUB-35936

Differential Revision: https://phabricator.twitter.biz/D310874