util-logging: Remove setting the user.dir property
Problem

When testing with JDK 11, we found that it is [[
https://bugs.java.com/bugdatabase/view_bug.do?bug_id=JDK-8202127 | no longer
possible ]] to set
the `user.dir` system property.

Solution

Remove setting the `user.dir` property in the test and adjust the test so that
it is testing the same behavior in a different way.

Result

The test passes when running with JDK 11.

JIRA Issues: CSL-8439

Differential Revision: https://phabricator.twitter.biz/D336895