util-tunables: Add an optional comment field to JsonTunable
Problem

Tunables are incredibly useful, however, it would be helpful to be able
to add human-readable context for each of the values in the JSON config.

Solution

Added a "comment" field that is ignored by the parser, but is readable
by humans.

Result:

Humans will be able to add comments to config.

Differential Revision: https://phabricator.twitter.biz/D353957