util-app: Address `Await` exceptions in app closing
Problem

There is the potential that the final `Await.ready` call in
`c.t.app.App#nonExitingMain` when closing emits a `TimeoutException`
which is swallowed by the current handling of all `NonFatal` exceptions.

In this case, swallowing the TimeoutException means that there's
no good signal on if the JVM should exit which potentially leaves
non-deamon threads running hanging the main thread of the application.

Solution

Make sure that we get signal when awaiting triggers a `TimeoutException`
in a way that can be handled appropriately by the caller of `nonExitingMain`,
which here is `App#main`.

Result

If we hang on awaiting during close which results in a TimeoutException, then
`exitOnError` will be properly invoked which should exit the JVM.

This is inspired by the work done in this PR and the conversation therein:
https://github.com/twitter/util/pull/251

JIRA Issues: CSL-8627

Differential Revision: https://phabricator.twitter.biz/D356846