util-core: Add undeclared dependency on scala-reflect.
Problem/Solution

The scala-reflect module is needed on the classpath to successfully
compile the util-core-util module. By default, Pants automatically
includes this module on the classpath so the lack of this dependency
hasn't been a problem so far. However, this dependency needs to be
explicitly declared in order for util-core-util to compile after
exporting the Pants build it to Bloop, see
https://github.com/pantsbuild/pants/pull/7961.

Differential Revision: https://phabricator.twitter.biz/D372472