util-core: Fix Reader.framed to work with non-List sequences
Problem

When using Reader.framed, if the framer doesn't emit a `List` the framer
will never emit any of that data. This is because it attempts to match
on the `::` class which is the cons cell of the type `List`.

Solution

Refactor the read loop to work for any `Seq`.

JIRA Issues: CSL-8868

Differential Revision: https://phabricator.twitter.biz/D378048