util-core: Add size to Memoize.Snappable interface
Problem:
The Memoize.Snapable interface has a 'snap' member
that creates a new map. For users just wanting to get the size
of the current table, creating a new map is unnecessary.

Solution:
Add a .size member to the interface. Store this in a simple variable
and increment it whenever something new is added to the memo table.

Differential Revision: https://phabricator.twitter.biz/D402887