util-core: Provide Future#toCompletableFuture
Problem

Java supplies a CompletableFuture out of the box since jdk8, which is used
heavily when integrating with Java libraries.  However, it's not always obvious
how to derive a Java CompletableFuture from a Twitter Future.

Solution

Provide a Future#toCompletableFuture method, analogous to Future#toJavaFuture,
which provides normal Twitter Future semantics, but via a CompletableFuture.
That means that canceling the returned CompletableFuture raises on the
underlying Future, and satisfying the underlying Future will satisfy the
returned CompletableFuture.

JIRA Issues: CSL-9098

Differential Revision: https://phabricator.twitter.biz/D408656