util-core: improve test coverage for method parse in Duration
Problem

The way it is currently implemented, the test will pass after the first
NumberFormatException for the string "". The rest of the strings are therefore
not tested.

Solution

Moving the assertion to within the foreach will ensure that all strings in the
Seq are tested.

Result

Improves code coverage in parse. Two additional branches in the code are now
checked.

Signed-off-by: Yufan Gong <yufang@twitter.com>

Differential Revision: https://phabricator.twitter.biz/D437860