util-jvm: Support for HotSpot Jvm that do not implement counters
Problem

When creating an com.twitter.jmv.Hotspot object on HotSpot JVMs that
do not implement JVM metrics, such as the HP NONSTOP JVM, the creation
fails with a javax.naming.OperationNotSupportedException.

Solution

As these metrics/counters are not mandatory, creation of the
com.twitter.jmv.Hotspot should just continue with a WARN message
that these counters are not available.

Result

Support for more Hotspot JVMs, such as HP NONSTOP JVM.