finagle-stats: Caches the JsonExporter's regex matching
Problem

finagle-stats doesn't cache which metrics JsonExporter matches against, which
may lead to reprocessing a lot of regex, if you have many metrics or a
complicated regex.

Solution

Cache the metrics that JsonExporter matches against, and provide a utility in
util-stats to make it easy to do such a thing.

JIRA Issues: CSL-9544

Differential Revision: https://phabricator.twitter.biz/D459391