util-routing: introduce experimental generic routing interfaces
Problem

Throughout the Twitter stack, we have multiple implementations
and representations of "routing" - taking an input and
searching for an associated destination. While the existing
implementations exist primarily at the `c.t.finagle.Service`
layer, the utility of a `Router` should not be limited to
a `c.t.finagle.Service`.

Solution

We introduce the `Router` and `RouterBuilder` traits that can
be used as a generic foundation for defining and building
routers.

JIRA Issues: CSL-8943

Differential Revision: https://phabricator.twitter.biz/D455757