util-core: Move PantsWorkaroundCore.scala to fix scoverage
Problem

In a previous commit the pants build was modifed to reduce noise and as
part of that the PantsWorkaroundCore.scala file was moved to a spot that
broke the sbt build. Specifically, the scoverage plugin isn't pleased to
find a source file outside of the perscribed directory structure.

Solution

Move it to a source directory.

JIRA Issues: CSL-9617

Differential Revision: https://phabricator.twitter.biz/D463341