util-core: Add Apply Method for Stream and Pool to InputStreamReader
Problem / Solution

`InputStreamReader` requires 3 parameters for construction, the input
stream, a chunk size, and a future pool for executing all I/O. There are
multiple scenarios where it makes sense to only provide the input stream
and the future pool, and use the `DefaultMaxBufferSize` as the chunk
size. `InputStreamReader` does not have an API which allows a user to do
that without explicitly specifying the chunk size. Let's add one.

Differential Revision: https://phabricator.twitter.biz/D480002