util-security: Move Credentials from util-core to util-security
Problem

The Credentials class exists in util-core, but a better place for it would be in util-security. It predates the creation of util-security by many years which is why it is where it is. Nothing within util-core depends on it though.

Solution

move it from util-core to util-security. the namespace should be changed from com.twitter.util to com.twitter.util.security.

JIRA Issues: CSL-9730

Differential Revision: https://phabricator.twitter.biz/D477984