util-app: simplify App close logic
Problem

App implements Closable and CloseAwaitably
in order to deal with some lifecycle edge
cases - primarily not related to App, but
Twitter Server. The close relationship is
not easy to understand.

Solution

We can simplify this relationship and keep
the existing behavior by instead
implementing `ClosableOnce`. We introduce
a new `CloseOnceAwaitably` that can wrap
a `ClosableOnce` and expose the same
underlying behavior as the previous
implementation without the need to track as
much state within App.

JIRA Issues: CSL-9864

Differential Revision: https://phabricator.twitter.biz/D497561