3rdparty: Upgrade scalatest to 3.1.2
Problem / Solution

We need to upgrade scalatest, so that we can stay up to date, take advantage of
new features, and test our libraries against dotty.

This is a more involved upgrade than normal.  A lot of old names have been removed,
and a bunch of namespaces were moved into new modules, and so we've had to move a lot
of things around.

TBR=true

NO-QUEUE=true

NO_USER_HOOK=1

Differential Revision: https://phabricator.twitter.biz/D518553