util-core: Rework Monitors Internals for Unidoc Generation
Problem / Solution

Trying to generate `unidoc` documentation for util fails when using
Scala 2.12.11. This is due to it not liking the `it` value of `Monitors`
which has type `Monitor$`. To work around the problem, we stepped the
version of Scala that we use go generate documentation down to 2.12.8.
However, it's not necessary for `Monitors` to have this value at all,
and so removing it is in our best interests moving forward.

JIRA Issues: CSL-9977

Differential Revision: https://phabricator.twitter.biz/D520046