util-core: Make Var.Observer public
Problem

The Var.Observer class is private[util], but a part of the abstract
`Var.observe` method which is intended to be public. This can be seen
through the `AbstractVar` class which is supposed to be an API for Java
users to make their own Var definitions.

Solution

Make it public (and final) so that it can be used.

Differential Revision: https://phabricator.twitter.biz/D520237