[util] Add Duration.fromHours and fromDays
Problem:

Java has TimeUnits for HOURS and DAYS, but ctu.Duration does not have the convenience methods to in it's fromXYZ suite.
Workarounds exist — `Duration.fromTimeUnit` or `DurationOps`, but it's still awkward to have the discrepancy in our API.

Solution:

I added `fromHours` and `fromDays` to the `Time` class, mimicking the `fromMinutes`, `fromSeconds`, etc methods.

JIRA Issues: CSL-7618

Differential Revision: https://phabricator.twitter.biz/D522734