util-stats: retain metadata for CumulativeGauges
Problem
-------
Metric Metadata is not retained properly when a gauge is constructed via CumulativeGauge's addGauge method.

Solution
--------
Remove the optimization that was leading to the lost MetricMetadata. The optimization was to avoid creating a new function on each call to whenGaugeNotPresent by having functions for each of the Verbosity levels. Now, there is more Metadata to store than simply Verbosity and we cannot have a cached function for all possible Metadata values.

Result
------
Metric Metadata is now properly populated and retained when CumulativeGauge's addGauge method.

JIRA Issues: CSL-9958

Differential Revision: https://phabricator.twitter.biz/D531151