util-app: Fix oss Scala 2.13 build
Problem:
The travis build failed to run `c.t.app.JavaFlagTest.testJavaEnumFlags` because
calling `Array.toSeq` is wrapping a Java array in `ArraySeq` in Scala 2.13, but
in pre Scala 2.13 world, it was wrapped in `WrappedArray`, which is deprecated
in Scala 2.13.

Solution:
Construct the error message with `Array.toList` instead of `Array.toSeq`
Parse the flag before reading from it.

JIRA Issues: CSL-10090

Differential Revision: https://phabricator.twitter.biz/D533992