finagle: configure SourceRole for finagle Clients and Servers
Problem
-------
Metric Metadata contains a field for the Role the service is acting as with respect to that metric, but nothing was populating it.

Solution
--------
Start populating the Client/Server role metadata in finagle.

Result
------
Client and Server related metrics created by finagle now have the role metadata populated.

JIRA Issues: CSL-9958

Differential Revision: https://phabricator.twitter.biz/D542596