Don't update the key everytime for Local.let
Problem:
Local.let always updates the key, even if it hasn't
actually changed. The current calls to `this.apply()`,
`this.update()`, and `this.set()` rely on ThreadLocal's
set and get methods, which look into map.

Solution:
Remove calls to `this.apply()`, `this.update()`,
and `this.set()`, and change key/values using Context
methods instead of ThreadLocal methods

JIRA Issues: CSL-6573

Differential Revision: https://phabricator.twitter.biz/D565448