util: Fix Scaladoc Issues for 2.11 with JDK 11
Problem / Solution

A bunch of Scaladoc changes were made recently in order to clean
up all of the warnings when building documentation for Util.
Unfortunately, when building that documentation using the combination of
Scala 2.11 with JDK 11, some of the changes result in an
"[error] bad constant pool tag x at byte y" error message, which
fails the build. This commit cleans up the offenders that have a problem
with the combination of Scala 2.11 and JDK 11.

JIRA Issues: CSL-10300

Differential Revision: https://phabricator.twitter.biz/D568686