finagle-internal/slo: Mark SLIs as key indicators in Metrics Metadata
Problem

The user-defined SLIs in SLO can be seen as more important metrics. And we want
to mark these metrics in the metrics metadata for future usage.

Solution

The `keyIndicator` field is already defined in the MetricBuilder and by default
the value is false.  If a metric is instrumented through SLO indicators, the
`keyIndicator` should be filled as true.  Introduce the
KeyIndicatorMarkingStatsReceiverProxy for this purpose.

JIRA Issues: CSL-9961

Differential Revision: https://phabricator.twitter.biz/D568839