util-stats: fix `InMemoryStatsReceiver` to handle printing floats in different locales
My laptop is set to use the Polish locale, which causes this test to fail locally. This is probably because in Polish, the thousands separator is `.` and the decimal separator is `,`, which is the reverse of English. (For example, in English, you write `1,234.56`, but in Polish, you write `1.234,56`.)

In particular, this assertion fails:

```
      assert(parts(5) == "g/h/i 1.000000")
```

because it's actually getting a value in the form `g/h/i 1,000000`.

Differential Revision: https://phabricator.twitter.biz/D566297