util-core: Activity.apply(Event) and Activity#stabilize should be safer
Problem

Activity.apply(Event) currently uses Var.apply(Event) to convert events to
vars.  We've seen problems with this approach in the past, so let's avoid it.
Activity#stabilize relies on Activity.apply(Event), so we should fix that too.

Solution

Activity.apply(Event) has a natural "no current known event" phase, which is
`Pending`!  We'll propagate witness registry to the underlying Event, and when
we're not watching, we'll fill in the gaps by describing it as being in
`Pending` state.

We'll also rewrite Activity#stabilize to not have to rely on Activity.apply(Event) anymore,
since it's incompatible with the new semantics.

JIRA Issues: CSL-10156

Differential Revision: https://phabricator.twitter.biz/D574843