util-benchmark: add unit, void, filter and map benchmarks
Problem:
We don't have benchmark tests for some common operators in ConstFuture. We also
don't benchmark `.map` and `.filter` for satisfying Promises. (There's `timeMap()`, but
it only accounts for the creation of a Promise and registering it on the wait
queue). This makes it hard to measure optimization changes.

Solution:
Add benchmarks for ConstFuture.unit, ConstFuture.void, ConstFuture.filter,
Promise.filter, Promise.map

JIRA Issues: CSL-10459

Differential Revision: https://phabricator.twitter.biz/D595080