util-*: Fix dependency to use scalatestplus mockito-3-3
## Problem

mockitoVersion is already 3.3.3, but scalatestplus currently incorrectly
uses mockito-3-2, which transitively depends on Mockito 3.2 as the name
suggests.

This seems to present no issue on Pants or sbt since Coursier would be
able to resolve to the later version of Mockito, and apparently
ScalaTestPlus is ok with having a different Mockito version on the
classapth.

However, this presents an issue for Bazel using bazel-multiversion
where we'd either have to choose between using the latest Mockito version in Source,
or declare that Mockito uses pvp versionScheme, at which point we would actually
have both Mockito 3.2 and Mockito 3.3 on the classpath.

## Solution

The solution is to upgrade scalatestplus to the correct mockito-3-3 module.

Differential Revision: https://phabricator.twitter.biz/D598104