util-logging: Add Formatter#formatMessageLines variant
Problem/Solution

There's currently no way to format the message lines on an already
processed message String. Update to add a `formatMessageLines`
variant which takes a (possibly already truncated and argument formatted)
message String and any Throwable from the LogRecord to create
the array of message lines.

Result

A way to pass in a differently formatted message String potentially
bypassing the `formatText` and truncation call in the version that
only takes a LogRecord.

Differential Revision: https://phabricator.twitter.biz/D619932