util: Make `ProxyByteReader` and `ProxyByteWriter` Public
Problem

`ProxyByteReader` is incredibly useful for reading specialized information out
of a `Buf` while still benefitting from all the already existing
goodness that comes for free from the `ByteReader` implementation.
However, it is not currently accessible to code outside of the
`com.twitter` namespace.

Solution

Make `ProxyByteReader` and `ProxyByteWriter` both public, and modify
`ProxyByteReader`s structure so that it aligns with `ProxyByteWriter`.

Differential Revision: https://phabricator.twitter.biz/D622705